+++ CHANGELOG.md
@@ -5,7 +5,11 @@
 
 [//]: # (Removed, Added, Changed, Fixed, Updated)
 
+### Fixed
+- fix bundled shrinker loading.
+
 ### Changed
+- and prefer bundled R8 if it’s newer by default.
 - change default values and logic for some configuration options for safety and convenience in the new projects (experimental and/or complicated options should be opt-in)
   - enableApiValidation = `false`
   - setupVerification = `false`